[TS-304] refactor bank functions?
