+++ libs/application/templates/social-insurance-administration/core/src/types.ts
@@ -1,3 +1,5 @@
+import { BankAccountType } from './constants'
+
 export interface BankInfo {
   bank?: string
   ledger?: string
@@ -13,4 +15,16 @@ export interface ApplicantInfo {
   bankAccount?: BankInfo
   phoneNumber?: string
   emailAddress?: string
+}
+
+export interface PaymentInfo {
+  bankAccountType: BankAccountType
+  bank?: string
+  ledger?: string
+  accountNumber?: string
+  iban?: string
+  swift?: string
+  bankName?: string
+  bankAddress?: string
+  currency?: string
 }
